+++ mrb_doc/shapes/rectangle.rb
@@ -24,10 +24,10 @@ end
 # Draw a rectangle with rotation
 # @param rectangle [Rectangle]
 # @param origin [Vector2]
-# @param angle [Float]
+# @param rotation [Float]
 # @param colour [Colour]
 # @return [nil]
-def draw_rectangle_pro(rectangle, origin, angle, colour)
+def draw_rectangle_pro(rectangle, origin, rotation, colour)
   # mrb_draw_rectangle_pro
   # src/mruby_integration/shapes/rectangle.cpp
   nil