+++ src/mruby_integration/core/drawing.cpp
@@ -1,5 +1,6 @@
 #include "raylib.h"
 #include "mruby.h"
+#include "mruby/compile.h"
 
 #include "mruby_integration/struct_types.hpp"
 
@@ -55,4 +56,24 @@ void append_core_drawing(mrb_state *mrb) {
   mrb_define_method(mrb, mrb->kernel_module, "end_mode2D", mrb_end_mode2D, MRB_ARGS_NONE());
   mrb_define_method(mrb, mrb->kernel_module, "begin_scissor_mode", mrb_begin_scissor_mode, MRB_ARGS_REQ(4));
   mrb_define_method(mrb, mrb->kernel_module, "end_scissor_mode", mrb_end_scissor_mode, MRB_ARGS_NONE());
+
+  mrb_load_string(mrb, R"(
+    def clear(colour: RAYWHITE)
+      clear_background(colour)
+    end
+
+    def drawing(&block)
+      begin_drawing
+      block.call
+    ensure
+      end_drawing
+    end
+
+    def scissor_mode(section, &block)
+      begin_scissor_mode(section.x, section.y, section.width, section.height)
+      block.call
+    ensure
+      end_scissor_mode
+    end
+  )");
 }