+++ test/fixtures/models.rb
@@ -1,0 +1,1 @@
+require 'test/fixtures/models/camera2d'