+++ test/fixtures/models/camera2d.rb
@@ -1,0 +1,31 @@
+def fixture_camera2d_drawing
+  w = RAYWHITE
+  r = RED
+  [
+    [
+      w, w, w, w, w, w, w, w, w, w,
+      w, w, w, w, w, w, w, w, w, w,
+      w, w, r, r, r, r, r, r, w, w,
+      w, w, r, r, r, r, r, r, w, w,
+      w, w, r, r, r, r, r, r, w, w,
+      w, w, r, r, r, r, r, r, w, w,
+      w, w, r, r, r, r, r, r, w, w,
+      w, w, r, r, r, r, r, r, w, w,
+      w, w, w, w, w, w, w, w, w, w,
+      w, w, w, w, w, w, w, w, w, w,
+    ],
+    [
+      r, r, r, r, r, r, w, w, w, w,
+      r, r, r, r, r, r, w, w, w, w,
+      r, r, r, r, r, r, w, w, w, w,
+      r, r, r, r, r, r, w, w, w, w,
+      r, r, r, r, r, r, w, w, w, w,
+      r, r, r, r, r, r, w, w, w, w,
+      w, w, w, w, w, w, w, w, w, w,
+      w, w, w, w, w, w, w, w, w, w,
+      w, w, w, w, w, w, w, w, w, w,
+      w, w, w, w, w, w, w, w, w, w,
+    ]
+  ]
+end
+